Refactor to use a factory inside the JSP
