+++ bundle/src/test/java/com/adobe/cq/components/oembed/OEmbedTest.java
@@ -11,6 +11,7 @@ import org.junit.Test;
 
 import com.adobe.cq.components.oembed.LinkFinder;
 import com.adobe.cq.components.oembed.OEmbedRenderer;
+import com.adobe.cq.components.oembed.LinkDiscoverer;
 import com.adobe.cq.components.oembed.OEmbedType;
 
 public class OEmbedTest {
@@ -19,7 +20,7 @@ public class OEmbedTest {
 	public void testPhoto() {
 		String endpoint = "http://www.flickr.com/services/oembed";
 		String url = "http://flickr.com/photos/bees/2362225867";
-		OEmbedRenderer renderer = new OEmbedRenderer();
+		LinkDiscoverer renderer = new OEmbedRenderer().getLinkDiscoverer();
 		renderer.fetchResponse(endpoint, url, 240, 180);
 		assertEquals(OEmbedType.PHOTO, renderer.getType());
 		assertEquals("https://farm4.staticflickr.com/3040/2362225867_4a87ab8baf_m.jpg", renderer.getURL());
@@ -30,7 +31,7 @@ public class OEmbedTest {
 	@Test
 	public void testDiscoverVideo() {
 		String url = "https://www.youtube.com/watch?v=9bZkp7q19f0";
-		OEmbedRenderer renderer = new OEmbedRenderer();
+		LinkDiscoverer renderer = new OEmbedRenderer().getLinkDiscoverer();
     assertTrue(renderer.discoverLink(url));
 		//renderer.fetchResponse(endpoint, url, null, null);
 		assertEquals(OEmbedType.VIDEO, renderer.getType());
@@ -40,7 +41,7 @@ public class OEmbedTest {
   @Test
   public void testDefaultFallback() {
     String url = "https://gfycat.com/WindyHiddenGemsbok";
-    OEmbedRenderer renderer = new OEmbedRenderer();
+    LinkDiscoverer renderer = new OEmbedRenderer().getLinkDiscoverer();
     assertTrue(renderer.discoverLink(url));
     assertEquals(OEmbedType.VIDEO, renderer.getType());
     System.out.println(renderer.getHTML());
@@ -49,7 +50,7 @@ public class OEmbedTest {
   @Test
   public void testFallback() {
     String url = "https://github.com/trieloff/github-oembed/blob/master/oembed.js";
-    OEmbedRenderer renderer = new OEmbedRenderer();
+    LinkDiscoverer renderer = new OEmbedRenderer().getLinkDiscoverer();
     assertTrue(renderer.fetchResponse("https://runtime-preview.adobe.io/github.com/trieloff/github-oembed/master/github-oembed/embed", url, null, null));
     assertEquals(OEmbedType.RICH, renderer.getType());
     System.out.println(renderer.getHTML());
@@ -58,8 +59,11 @@ public class OEmbedTest {
   @Test
   public void testDefaultDiscovery() {
     String url = "https://gist.github.com/trieloff/013509c40db9860746fe3977acadb676";
-    OEmbedRenderer renderer = new OEmbedRenderer();
-    renderer.setOEmbedEndpoints(Arrays.asList(new String[] {"https://runtime-preview.adobe.io/github.com/trieloff/github-oembed/master/github-oembed/embed"}));
+    OEmbedRenderer r = new OEmbedRenderer();
+    r.setOEmbedEndpoints(Arrays.asList(new String[] {"https://runtime-preview.adobe.io/github.com/trieloff/github-oembed/master/github-oembed/embed"}));
+    
+    LinkDiscoverer renderer = r.getLinkDiscoverer();
+    
     assertTrue(renderer.discoverLink(url));
     assertEquals(OEmbedType.RICH, renderer.getType());
   }
@@ -67,8 +71,11 @@ public class OEmbedTest {
   @Test
   public void testDefaultDiscoveryGist() {
     String url = "https://github.com/trieloff/github-oembed/blob/master/oembed.js";
-    OEmbedRenderer renderer = new OEmbedRenderer();
-    renderer.setOEmbedEndpoints(Arrays.asList(new String[] {"https://runtime-preview.adobe.io/github.com/trieloff/github-oembed/master/github-oembed/embed"}));
+    OEmbedRenderer r = new OEmbedRenderer();
+    r.setOEmbedEndpoints(Arrays.asList(new String[] {"https://runtime-preview.adobe.io/github.com/trieloff/github-oembed/master/github-oembed/embed"}));
+    
+    LinkDiscoverer renderer = r.getLinkDiscoverer();
+    
     assertTrue(renderer.discoverLink(url));
     assertEquals(OEmbedType.RICH, renderer.getType());
   }
@@ -90,7 +97,7 @@ public class OEmbedTest {
   @Test
   public void testTwitterCardsMedium() throws Exception {
     String url = "https://medium.com/adobe-io/what-event-based-architectures-mean-for-application-integration-662932169d28";
-    OEmbedRenderer renderer = new OEmbedRenderer();
+    LinkDiscoverer renderer = new OEmbedRenderer().getLinkDiscoverer();
     assertTrue(renderer.discoverLink(url));
     assertEquals(OEmbedType.CARD, renderer.getType());
   }
@@ -99,7 +106,7 @@ public class OEmbedTest {
   @Test
   public void testTwitterCards() throws Exception {
     String url = "http://www.spiegel.de/wissenschaft/natur/great-barrier-reef-zwei-drittel-der-korallen-drohen-abzusterben-a-1142645.html";
-    OEmbedRenderer renderer = new OEmbedRenderer();
+    LinkDiscoverer renderer = new OEmbedRenderer().getLinkDiscoverer();
     assertTrue(renderer.discoverLink(url));
     assertEquals(OEmbedType.CARD, renderer.getType());
   }
@@ -107,7 +114,7 @@ public class OEmbedTest {
 	@Test
 	public void testDiscoverPhoto() {
 		String url = "http://flickr.com/photos/bees/2362225867";
-		OEmbedRenderer renderer = new OEmbedRenderer();
+		LinkDiscoverer renderer = new OEmbedRenderer().getLinkDiscoverer();
 		assertTrue(renderer.discoverLink(url));
 		assertEquals(OEmbedType.PHOTO, renderer.getType());
 		assertEquals("https://farm4.staticflickr.com/3040/2362225867_4a87ab8baf_b.jpg", renderer.getURL());
